Q4 final bugfix

+++ homework/utils.py
@@ -223,7 +223,7 @@ class DepthError(object):
         # rmse_log = (np.log(self.gt) - np.log(self.pred)) ** 2
         # rmse_log = np.sqrt(rmse_log.mean())
 
-        abs_rel = np.mean((np.abs(np.array(self.gt) - np.array(self.pred)) / (np.array(self.gt) + 1e-10)).float())
+        abs_rel = np.mean((np.abs(np.array(self.gt) - np.array(self.pred)) / (np.array(self.gt) + 1e-10)))
 
         # sq_rel = np.mean(((self.gt - self.pred) ** 2) / self.gt)
